+++ emle/_resources.py
