+++ emle/calculator.py
@@ -287,7 +287,7 @@ class EMLECalculator:
             the calculator.
         """
 
-        from ._utils import _fetch_resources
+        from ._resources import _fetch_resources
 
         # Fetch or update the resources.
         _fetch_resources()
